+++ drivers.c
@@ -859,8 +859,8 @@ static const struct gps_type_t rtcm104v2 = {
 
 static gps_mask_t rtcm104v3_analyze(struct gps_device_t *session)
 {
-    uint length = getbeuw(session->packet.inbuffer, 1);
-    uint type = getbeuw(session->packet.inbuffer, 3) >> 4;
+    uint16_t length = getbeuw(session->packet.inbuffer, 1);
+    uint16_t type = getbeuw(session->packet.inbuffer, 3) >> 4;
 
     gpsd_report(LOG_RAW, "RTCM 3.x packet type %d length %d words: %s\n",
 	type, length, gpsd_hexdump_wrapper(session->packet.inbuffer,
